+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - Component must have domain and it's participants must be in the same domain
 - Change tuc comment back to `//` from `#`, so it won't conflict with `#colorHash` in notes, etc.
 - Transform *italic* in tuc files
+- Fix Read/Post event by FQ name for events with only one case
 
 ## 0.2.0 - 2020-09-01
 - Parse left and right notes in tuc